add interfaces; parse interfaces first
